make duration a floating point number
git-svn-id: svn+ssh://kobol/srv/svn/jack_oscrolloscope@11 8b2478be-e929-0410-a59d-bb23318da49a

+++ main.h
@@ -25,18 +25,18 @@
 #define DEFAULT_DURATION            5
 
 
-extern bool g_run;
-extern int  g_nports;
+extern bool     g_run;
+extern int      g_nports;
 
-extern int  g_ticks_per_frame;
-extern bool g_scrolling;
-extern int  g_width;
-extern int  g_height;
-extern int  g_total_height;
-extern bool g_use_gl;
+extern int      g_ticks_per_frame;
+extern bool     g_scrolling;
+extern int      g_width;
+extern int      g_height;
+extern int      g_total_height;
+extern bool     g_use_gl;
 
-extern int  g_duration;
-extern bool g_show_clipping;
+extern float    g_duration;
+extern bool     g_show_clipping;
 
 extern Uint32  *g_colors;
 extern float   *g_scales;
